refactor: move all type definitions into single `typeDefs.ts` file
Closes #56
