--- src/modules/userGroups/userGroupsTypes.ts
@@ -1,9 +1,0 @@
-import type { BaseEntityModel } from "~/modules/allEntities/allEntitiesTypes";
-
-export interface UserGroupModel extends BaseEntityModel {
-  id: string;
-  name: string;
-  userIds: string[];
-}
-
-export type UserGroupsByIdModel = Record<string, UserGroupModel>;
